Replace most uses of {const_}iterator by 'auto'
This is some typical code pattern we _often_ used in openmsx:

* in the .hh file:
	typedef std::map<std::string, unsigned> MyMap;
	MyMap myMap;
* in the .cc file:
	MyMap::iterator it = myMap.find(str);
	...
	for (MyMap::const_iterator it = myMap.begin();
	     it != myMap.end(); ++it) { ... }

Note that without using the typedef, both lines in the .cc file would still be
longer. By using the c++11 'auto' feature, this example can be rewritten like:

	auto it = myMap.find(str);
	...
	for (auto it = myMap.begin(); it != myMap.end(); ++it) { ... }

And we don't need the typedef anymore.

--

Actually in c++11, the for-loop can still be further simplified:
	for (auto& element : myMap) { ... }
Though this syntax is not yet supported in visual studio 2010 (it is already
supported in 2012, or at the very least in the next version to be release later
this year).

--

There are still many more opportunities to simplify the code by using auto,
this patch only searched for code sequences using iterator or const_iterator.

Note that I simplified both iterator/const_iterator using the same pattern (so
the resulting type is now sometimes iterator while it used to be
const_iterator). If I really wanted I could have forced a const_iterator by
using cbegin() and cend(), but I didn't bother, also because hopefully soon we
can rewrite the loops using the new for-loop syntax.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12962 7512d703-071c-0410-b5c2-d2bdd895b7d1