Replace {string,string_ref}::size_type with auto
This patch replaces the pattern:
	string::size_type pos = myString.find("foobar");
with:
	auto pos = myString.find("foobar");

The latter is shorter. But more important, this code has already been the
source of bugs twice (though this patch doesn't solve any more bugs):

1) In the past we sometimes assigned the result of the find() method to an
'unsigned' variable. This works fine on 32 bit systems, but it's a bug on 64-bit
systems. Certainly when this result is later compared with string::npos.

2) For efficiency we sometimes convert std::string into string_ref. The latter
has a very similar interface as string. Though the string_ref::find() method
returns the type string_ref::size_type instead of string::size_type. At the
moment both types are different on a 64-bit platform: 32bit vs 64bit (though
this will change in the future).

By using 'auto' you get the same as the actual return type from the find()
function. So you don't have to explicitly/redundantly specify the type, with
the risk of getting it wrong and forcing a conversion.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12984 7512d703-071c-0410-b5c2-d2bdd895b7d1