[2/4] Implement input layer system: c++ part
As already mentioned in the previous patch, the input layer system is a cleaner
solution for the hack we introduced in revision 12986. The problem that this
hack solved was the following:

   Revision 12986 introduced the 'OSDControl events', these are abstract events
   that will typically be used to control OSD elements (like the OSD keyboard
   or menu). These events e.g. aggregate the cursor and all joystick movement
   events into abstract movement events. So the OSD widgets can use these
   abstract events instead of having to listen to all possible movement events
   (and similar for action buttons).

   There was however one problem with this approach: for example the OSD
   keyboard widget is now bound to 'OSDControl up/down/left/right' events
   instead of (as before) directly to the cursor key events. So the cursor key
   events are no longer bound. A consequence of this is that these events are
   also no longer blocked from the MSX (the HotKey class doesn't 'consume' the
   event anymore). So while using the OSD keyboard (with the cursor keys), the
   MSX also receives events for those keys, making entering text with the OSD
   keyboard impossible.

   As a temporary hack, we blocked all events to the emulated MSX as soon as
   at least one OSDControl event was bound (so e.g. when the OSD keyboard was
   active).

The above hack works fine for the combination OSD-keyboard/emulated-MSX, but
more general, when there are multiple OSD widgets, those widgets may also want
to block events from each other. Or in other words, at some point a widget may
want to have (semi-)exclusive access to the events.

This patch series implements an 'input layer system'. Tcl scripts can now
instead of globally binding commands to events (e.g. hotkeys) bind events to a
specific input layer. Later such a layer can be activated (and deactivated). An
active layer can be in blocking or non-blocking mode. Blocking mode means that
even if an event is not used by the current layer, the event will not be passed
to the lower layer or to the emulated MSX (it will still be passed to the
globally bound hotkeys).

Very briefly, the (un)bind command now takes an optional -layer option. And there
are two new commands {de}activate_input_layer to (de)activate the layers, the
activate command takes an optional -blocking flag. See next patch for a more
detailed description of the commands.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13000 7512d703-071c-0410-b5c2-d2bdd895b7d1