Use c++11 new for-loop syntax
Now that visual studio 2012 has support for windows XP (again) we can start
using this feature.

A pattern like:
	auto& pluggables = pluggingController.getPluggables();
	for (auto it = pluggables.begin(); it != pluggables.end(); ++it) {
		Pluggable& pluggable = *it;
		...
	}
can be replaced with
	for (auto& pluggable : pluggingController.getPluggables()) {
		...
	}

In general the new for loop can iterator over anything that has a std::begin()
and a std::end(). So most often this will be STL containers, but it can also
be plain arrays or user defined types.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13016 7512d703-071c-0410-b5c2-d2bdd895b7d1