Initialize std::string from string_ref cleanup
To construct/initialize a std::string from a string_ref we used to do this:
   my_string(my_string_ref.data(), my_string_ref.size())
instead of
   my_string(my_string_ref.str())

The idea was that in the first case the std::string is directly constructed
from the correct data (the string_ref data() and size() methods are inlined
and are very cheap), so only one copy of the data. While in the second case
we first construct a temporary string object and initialize the 'my_string'
object with that, so potentially two copies of the strinf data.

In c++98 the first case may indeed have been better (though I'm not 100% sure,
copy-elision might have been able to avoid one of the two copies). But in c++11
we can move construct 'my_string' from the temporary, so there's certainly no
extra copy.

I inspected the generated code and the second case is better in c++11. It's
also easier/shorter in the source code. So this patch changes the code to use
the second pattern.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13023 7512d703-071c-0410-b5c2-d2bdd895b7d1