Fixed serialization bug in V9990 code
In V9990CmdEngine we forgot to serialize the 'clock' member variable. This
patch fixes that.

At the same time I copied an optimization over from the VDP (V99x8) code,
because it changes the way in which this member should be serialized: instead
of keeping a Clock object and increasing that using a number of cycles, we keep
a EmuTime and increase that with the duration of those cycles. This removes
a multiplication from the inner loops in the command engine. See revision
12401 for more details.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13045 7512d703-071c-0410-b5c2-d2bdd895b7d1