Address question in revision 13051
(This patch also changes some code layout, but since this didn't interfere with
the main change of this patch, I didn't bother splitting the patch in two).

This patch is a response to the following question in the code:
  toBeRescheduledEvents.clear(); // Do I need to clear this local vector,
     // before it goes out of scope? Or will C++ invoke the destructor and
     // will the destructor handle it?
In C++ when an object goes out of scope (a local variable in this case), the
destructor of that object indeed gets called. The destructor of a std::vector
will in turn destruct all the elements it currently contains. So this explicit
clear() is not needed.

(detail)
Right above this line there was a loop that copied the content of one vector
to the back of another vector:
	for (auto& e : toBeRescheduledEvents) {
		toBeScheduledEvents.push_back(e);
	}
A slightly faster alternative is to use the range insert() method on
std::vector. (Insert at the end of toBeScheduledEvents, the range from begin to
end of toBeRescheduledEvents).
	toBeScheduledEvents.insert(
		toBeScheduledEvents.end(),
		toBeRescheduledEvents.begin(), toBeRescheduledEvents.end());

In c++11 there's an even better alternative: 'move' instead of 'copy'. Making a
copy of a shared_ptr is not that expensive but also not trivially cheap
(internally it has a reference counter that needs to be manipulated using
atomic operations (for thread-safety), atomic operations are more expensive
than regular operations on multi-processor systems). The code above makes a
copy of the events in the vector and immediately after destroys those
copied-from objects. This is a perfect scenario to use moves:
	move(toBeRescheduledEvents.begin(), toBeRescheduledEvents.end(),
	     back_inserter(toBeScheduledEvents));

Very briefly explained: you can move instead of copy when you know you won't
need the source object anymore. This allows the implementation of the move
operation to 'steal the internal guts' of the source object. The only
requirement is that the moved-from object is left in a destructible or
assignable state. So specifically for a shared_ptr this means that a move
operation will assign the internals from the source to the destination (without
manipulating the reference count) and set the source to nullptr.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13058 7512d703-071c-0410-b5c2-d2bdd895b7d1