Replace most uses of std::set<> with std::vector<>
A set is a useful data structure, though in openMSX we overused it a bit. Or
sometimes we conceptually needed a set, but it was more efficiently implemented
using a vector. Some examples:

We use a std::set<> (very often a set of strings) when we want
 - the items in sorted order
 - make sure the items are all unique
 - quick access to the biggest element (looser variant of the 1st item)

The most often used interface in the openMSX code that required a set<string>
was the tabcompletion code. In some case it needs to show the list of all
possible completions in sorted order. Also internally it's useful to have fast
access to the shortest partial match, this is easy in a sorted data structure.

So because internally it's useful to have (the subset of partial matches) in
sorted order we required to already pass the whole data structure in sorted
order. Almost all users of the tabcompletion functions (used to) dynamically
construct the set of possible completions. So they in a loop insert the items
one-by-one in a set<string>, this is a O(n*log(n)) operation. After this patch
(most of) those callers instead create a vector<string> (this is O(n)) and the
tabcompletion routine will internally sort (a subset of) that vector (this is
O(m*log(m)), with m<=n). So this has the same theoretical time complexity as
before, but because operations on vector are mostly a lot faster than
operations on set, the whole approach may actually be faster.

I said that most callers now pass a vector<string>, this is not completely
true. A few still pass set<string>, many pass vector<string>, but still many
other pass vector<string_ref> or even an array of string-literals. Any
collection of string-like objects can now be passed (the tabcompletion routines
are now templatized). vector<string_ref> has the advantage that the strings
don't need to be copied (e.g. because those same strings are already present in
some other data structure, it's cheaper to just refer to them). The array of
string literals is even cheaper, it doesn't require any runtime setup, the
whole structure just sits already in memory. This is all good for speed (e.g.
less heap memory allocation, better cache locality).

Though speed is not the main motivation for this change (tab completion is only
used interactively and it was already more than fast enough). The reason for
this change is reduced code size. Both reduced executable size (binary is about
16kB smaller). But also less and or simpler source code. E.g. the source code
to build an array of string literals is a lot more compact than to build a
set<string>.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13075 7512d703-071c-0410-b5c2-d2bdd895b7d1