Added xrange utility
This utility allows to rewrite a loop like
	for (int i = 0; i < 100; ++i) { .. }
as
	for (int i : xrange(100)) { .. }
So it's only syntactic sugar. Both generate exactly the same binary code.

The main motivation for adding this is in combination with other functions. So
for example
	unsigned size = myStringRef.size();
	for (unsigned i = 0; i < size; ++i) { .. }
can be rewritten as
	for (auto i : xrange(myStringRef.size()) { .. }

The main advantage of this alternative (beside it being slightly shorter) is
that the code is now agnostic of the actual return type of the size() method.
If that method returns a 32-bit value, the type of the induction variable will
also be 32-bit (and with the same signedness), if the return type was 64-bit,
the induction variable type is also 64-bit. In later patches we will change
some return types from 32->64 bit.

--

The following patches will add 'better' 64-bit support to openMSX. openMSX
already compiles and runs fine on both 32 and 64 bit systems. But internally in
many places we use 32 bit variables for things that could in theory be larger.
The most obvious such thing is the size of a file.

The first few patches in this series will adapt some of the basic helper
functions/classes. Later patches change the code that builds on top of those
helpers.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13077 7512d703-071c-0410-b5c2-d2bdd895b7d1