64-bit string_ref
string_ref is a class that refers to a string, it does not own the memory that
holds the string content. It has a pointer to the beginning of the string and a
variable that indicates the size of the string.

This patch changes the size variable from 32 to 64 bit (on 64 bit systems).

We most likely won't need such big strings. Though changing this doesn't have
any measurable cost. And it does allow for some code cleanups. For example
visual studio (rightfully) generates warnings when a 64-bit value is truncated
to 32-bit. This e.g. happens when we (used to) construct a string_ref from a
std::string object. To silence such warnings we had to add extra casts in the
code, so those can be removed now.

Also it just feels better to not artificially limit helper functions to 32 bit
on 64 bit systems. You never know that in the future you may need such a helper
on a very large data structure (e.g. a large harddisk image, or a laserdisc
video).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13078 7512d703-071c-0410-b5c2-d2bdd895b7d1