64-bit sha1sum
When calculating the sha1sum of a block of memory, the size of that block was
limited to 32 bit. This patch changes it to 64-bit (on 64-bit platforms).

When later we add support for large harddisk images, we *might* actually have
to compute such large sha1sums. (Actually we'll likely incrementally compute
the sha1sum on many small subblocks, but there's no good reason to already
limit the possibility of using a large block at this level).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13080 7512d703-071c-0410-b5c2-d2bdd895b7d1