64-bit PatchInterface
We have an abstract 'PatchInterface' class that efficiently allows to have a
patched view on a file without having to load that whole file in memory (e.g.
this is useful for patched hardisk images and to a lesser degree also for
diskimages).

This patch extends PatchInterface to work with 64-bit underlying sizes.
Although the only concrete patch fileformat (IPS) that we currently support is
limited to 24-bit offsets.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13087 7512d703-071c-0410-b5c2-d2bdd895b7d1