Fixed compile error on windows, related to ssize_t
It seem the type 'ssize_t' is not a type defined by the C standard (it is a
POSIX extension). So there's no standard header file that can be included
to have this type defined. It seems that in the DirAsDSK code on windows the
ssize_t type was known, but not in the Base64 code (on linux compilation went
file for both).

Luckily it wasn't too hard to rewrite the code to not need ssize_t.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13093 7512d703-071c-0410-b5c2-d2bdd895b7d1