Added KeyRange utility
Loosely inspired by boost::adaptors::map_keys.

This utility takes as input a map (std::map or StringMap) and provides a
alternative view on this data structure that only shows the keys in the map.
For example:
	map<string, int> myMap = ...;
	for (string& s : keys(myMap)) {
		cout << s << endl;
	}
This will print all the keys from the 'myMap' map. Note that this will not
construct a temporary collection with all those keys. Instead the keys are
(lazily) extracted from the underlying map.

In this patch I use this utility to simplify some tab-completion code: We used
to construct a temporary vector<string_ref> collection that contained all the
keys of a StringMap. Now we can simply pass the key-view of the map to the
completion routine.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13101 7512d703-071c-0410-b5c2-d2bdd895b7d1