Added array_ref utility
This is similar to the string_ref utility (both are modelled after utilities
with the same that are proposed for the next c++ standard). An array_ref<T>
refers to an array of T objects, but it does not own that array. Internally it
has just a pointer to the first object in the array plus a value to indicate
the length of the array.

The methods getLevelStrings() and getUpdateStrings() in the CliComm class
conceptually return ia collection of string-literals. An efficient way to
implement this in c/c++ is to return a pointer to a c-style array,
unfortunately this looses the information about the size of the array. This
means that the result of these methods can't directly be used to pass to other
functions that expect a range (this needs the begin and end of the array). This
patch changes the return type of the two methods to array_ref<const char*>.

We can use this to avoid creating an extra temporary collection to pass to the
tab-completion routines.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13102 7512d703-071c-0410-b5c2-d2bdd895b7d1