Don't crash on loading a manually edited savestate
We tried to load a replay, but we didn't have a cas file that was used in this
replay. We manually edited the replay (removed the file), but when loading this
new replay, openMSX would crash.

The reason is that this violates an internal invariant: the cassetteplayer was
in 'PLAY' mode according to the savestate, but the code assumes this is only
possible when there's actually a cassette image inserted.

Fixed by forcing the cassetteplayer in 'STOP' state when there was no cassette
image loaded.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13104 7512d703-071c-0410-b5c2-d2bdd895b7d1