get_screen_mode should not give an error on an invalid screen mode
Vampier reported that occasionally the info_panel would stop with the error
  Illegal VDP mode!
After some investigation we found that the 'get_screen_mode' Tcl proc generates
this error (the info panel uses this proc to show the current screen mode).
It's much more convenient if a query proc like get_screen_mode always
successfully returns a result. And a (successful) result could also be 'invalid
screen mode'.

So that's what this patch does. I verified that the info panel now shows the text
'invalid' when the VDP is configured with invalid mode selection bits.

BTW: setting a new screen mode requires to change 2 VDP registers, this cannot
be done atomically, so when the get_screen_mode proc is used in the (brief)
period when only one of the two registered is changed yet, it's possible you
read an invalid screen mode (thus even if the MSX program never intentionally
sets an invalid mode).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13107 7512d703-071c-0410-b5c2-d2bdd895b7d1