Fixed interference between console and osd menu
When you open the openMSX console while the OSD menu was already open, then key
presses (e.g. cursor up/down) would get processed by both the console and the
menu. This is a regression since revision 12987. This revision changed the OSD
menu script to bind to the new OSDControl events instead of directly to
keyboard events (this makes the menu also controllable via joystick, to make it
usable on Android).

The console has the highest priority in processing events. If it handles (or
IOW consumes) an event, that event is no longer send to the lower priority
listeners (e.g. to Tcl scripts that bind to events). Though the console only
consumes keyboard events, no OSDControl events. Hence the regression.

Shortly after creating the OSDControl events, we introduced the input layer
system. This system prevents different Tcl scripts that all listen to
OSDControl events from interfering with each other: Tcl scripts bind to a
specific layer and layers can be stacked and block events from passing to each
other.

This patch makes the openMSX console also use this input layer system. It adds
a small piece of Tcl that (de)activates a new input layer whenever the console
is (de)activated. Even though the rest of the console is not implemented in
Tcl.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13116 7512d703-071c-0410-b5c2-d2bdd895b7d1