Fixed potential assert in OSDConsoleRenderer
This was a logic error in the console rendered-text-cache. Though it's very
hard to trigger this assert: there had to be more text elements on the screen
than the size of cache (currently 250). And the least recently shown text
element had to be the last element from the cache.

Manuel reported this assert (with backtrace). I tried but I couldn't reproduce
it. I believe this patch fixes the problem, but I couldn't verify it.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13132 7512d703-071c-0410-b5c2-d2bdd895b7d1