Fixed a too strict assert
This fixes Debian bug:
  Bug#698315: openmsx: FTBFS due to invalid alignment assumptions

On platforms like m68k, the type 'uint32_t' only has alignment 2 (instead of 4
like on most other platforms). The code had an assert that e.g. the class B32
(which internally uses a uint32_t) was _exactly_ 4 bytes aligned. Though this
is too strict (and is not true on m68k), now we instead check that the
alignment is _at_most_ 4 bytes.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13141 7512d703-071c-0410-b5c2-d2bdd895b7d1