Fixed(?) filename tabcompletion
The code relied on internal pointers to string objects remaining valid after
those strings are copied (or moved) to a different object. Apparently this
happened to work on gcc/Linux, but on vc++/windows it didn't. According to the
c++ standard this is indeed undefined behavior.

This should solve bug
   [3603207] tab completion bug
Though apparently the bug only manifests itself on windows, so I couldn't
actually test that it's really solved.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13151 7512d703-071c-0410-b5c2-d2bdd895b7d1