Made all events printable/comparable (fixes crash in debug build)
On my laptop (but not on my desktop) a debug build of openMSX would crash soon
after startup. The reason was a debug print in InputEventGenerator::handle().
It tried to print an Event object that represented a SDL_VIDEOEXPOSE event.

Before this patch not all Event subclasses had a _working_ Event::toString()
method (and similar for the <, ==, != operators). This patch does add working
implementations.

The design was a bit strange: the base Event class offered a toString() method,
but some subclasses had an implementation that would do the equivalent of
'assert(false)'. I've replaced those with real implementations. An alternative
could have been to introduce an intermediate base class and move the toString()
method to that level. (Though for debugging it's useful to be able to print all
Events.)

Note that while all Events now have a working toString(), it's not yet possible
to reconstruct all events from such a string representation.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13157 7512d703-071c-0410-b5c2-d2bdd895b7d1