Avoid memory reallocation in OutputBuffer
During profiling I noticed there was quite some memcpy'ing going on. This came
from the realloc() function call in the OutputBuffer class. This class is used
to store the result of an in-memory savestate (so happens regularly when
reverse is enabled).

Before this patch, the OutputBuffer class started with an initial buffer size
of zero. When data gets added, we need to (re)allocate bigger and bigger
buffers. A reallocation has to allocate more memory, but potentially also copy
the existing data to the new location (realloc() tries to grow the buffer
in-place, but that's certainly not always possible).

This patch tries to estimate the required buffer size (based on previous buffer
sizes) and then immediately allocate a buffer of that size. So that
reallocations are less frequently needed.

The code already made the buffer size grow exponentially. That means that on
average each byte in the buffer only needs to be copied once. Still, this copy
step was visible in the top 10 of a profile-run. The specific test I ran was
zexall without video/sound rendering. On my desktop Linux box, this test is now
1%-2% faster. I'm not sure how much the speed difference will be in other
scenarios.

On embedded devices it does make a noticeable speed difference when reverse is
enabled or not. Hopefully this patch makes the speed-penalty now a bit smaller.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13161 7512d703-071c-0410-b5c2-d2bdd895b7d1