Tweaked CPUClock::enableLimit()
I noticed that the CPUClock::enableLimit(bool) method was quite high in a
profile result. Furthermore this method was always called with a constant
boolean parameter (I mean either with the constant 'false' or 'true', but never
with a runtime variable boolean value). The implementation of this method was
(mostly) different depending on which parameter value was passed. So now I've
split this method in two: a enableLimit() and a disableLimit() method, both
without parameters.

The new enableLimit() method was only called from one location. The new
disableLimit() method was called rom two locations, but the implementation of
disableLimit() is extremely simple. So I've also made both methods inline.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13162 7512d703-071c-0410-b5c2-d2bdd895b7d1