Reduced code bloat in Z80 read/write memory routines
The R800 has a limited form of page-mode RAM access. That is if the high byte of
the address has not changed compared to the previous access, the row address
(RAS) of the RAM chips does not have to be reselected. This can safe a clock
cycle. It's limited in the way that not always when the high byte remains the
same, this faster page-access takes place. The exact conditions on when or when
not this optimization can happen are not so easy, but in some instructions you
know when this cannot happen and you can statically insert this extra cycle
(instead of dynamically checking for it).

So to gain some emulation speed, most of the memory access functions were
templatized with pre- and or post-page-break booleans that indicate whether
there's certainly a page break or when to check for it. These template
parameters are used to call other templatized helper functions that in the R800
code do or do not add the extra cycle or perform a dynamically check. The Z80
does not have such page-breaks, however to keep the emulation code as much as
possible the same between Z80 and R800, those annotations are still present for
Z80, but the implementation of the helper functions is empty. This generates
optimal code for both Z80 and R800.

So for example the method
  template<bool PRE_PB, bool POST_PB> byte CPUCore<T>::RDMEMslow(...)
generates, depending on the PRE_PB and POST_PB booleans, for R800 4 slightly
different versions. For Z80 all 4 combinations generate the same code with all
the page-break stuff nicely optimized away.

So the generated code is perfectly fine, the only downside is that there are 4
identical routines for Z80. This patch addresses this code bloat. It inserts an
extra helper routine in front of the original templatized routine that in case
of R800 simply delegates to the original routine. But in case of Z80, it changes
the 2 boolean template parameters to false before calling the original. This
extra helper function can be completely optimized away.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13163 7512d703-071c-0410-b5c2-d2bdd895b7d1