Small tweak in Scheduler::scheduleHelper()
Part of this method dealt with the case that the queue of SynchronizationPoints
is empty, but other parts of this method (already) assumed this queue can never
be empty. This assumption (queue never empty) is indeed true.

This patch thus removes the first check. This is only a very minor
optimization. Though this method was in the top-15 of some profile runs, after
this change it dropped a few positions. Still I doubt this patch on it's own
will make any noticable difference. but many small tweaks like this combined
may.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13164 7512d703-071c-0410-b5c2-d2bdd895b7d1