Don't try to compress small blocks for in-memory snapshots
I found that the serialization code for the RP5C01 (the RTC) registers tried to
compress a block of 52 (4x13) bytes. But often the result was bigger than the
input (typically around 64 bytes + 8 bytes to store the length of the
compressed block). This is not too bad, except that trying to compress this
block also takes a relatively long amount of time. This is because internally
the snappy compression algorithm uses a hashtable and this table needs to be
cleared first. This setup cost is the same for very small and for very big data
blocks.

So when also taking this setup time into account, it's probably better to just
not try to compress small memory blocks.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13167 7512d703-071c-0410-b5c2-d2bdd895b7d1