Fixed initial state of 'disallow{Read,Write}Cache
This fixes a regression recently introduced in revision 13165: Manuel noticed
that one of his replays desynced after this revision.

Very strictly speaking the bug was not introduced in rev13165, but it did
uncover a much older bug. Some background info:

  Memory location 0xffff in the Z80 address space contains the secondary slot
  select register. More specifically depending on which primary slot is
  selected in page 3 (0xc000-0xffff) it contains one of the possibly 4
  secondary slot registers. However if some slot is not expanded, then location
  0xffff is just a regular memory location.

  So when location 0xffff does contain a subslot register, the cpu-cache-line
  mechanism must be disabled for that region. This is done via the
  disabledReadCache and disabledWriteCache variables.

Before revision 13165, writing to the primary slot select register would always
recalculate the disabled{Read,Write}Cache variables. But rev13165 introduced an
optimization that only recalculates this state when it actually changes.

This patch should normally be fine, except for the fact that we had a bug in
our code: we forgot to correctly initialize these variables. So before the
primary slot register was written for the first time, these variables could be
wrong (and they were also wrong right after you remove a slotexpander
at-run-time, though this rarely happens).

But after rev13165 we update the disabled{Read,Write}Cache variables much less
frequently, so the wrong initial state existed for much longer. After we know
what caused the problem, it's easy to fix: just properly initialize it.

Ok, but even before rev13165 the initial state was wrong ... Why didn't we
notice this? I guess this is because all MSX BIOS initialization routines first
write the primary slot register before doing something with the secondary slot
register. So technically it was a bug, but it could never cause any harm.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13169 7512d703-071c-0410-b5c2-d2bdd895b7d1