Fixed mistake in EmuTime variable-type.
Bug reported by BiFi. This was a regression introduced in revision 13164. This
bug only manifested itself on 32-bit platforms (thanks to Quibus for this
suggestion).

Some background info:
 Emutime is a 64bit value. These types are often simply passed around (rather
 than actually used by the function). On 32-bit systems it's more efficient to
 pass a pointer to such a 64-bit value than to pass the actual value (this only
 takes 1 (scarce) register instead of 2). While on 64-bit systems it's cheaper
 to directly pass the value around. To choose between these two alternatives we
 created a EmuTime::param typedef that depending on the platform chooses one of
 the two passing mechanisms.

The bug was that in the code we wrongly used Emutime::param instead of simply
EmuTime because in this case we really wanted a copy. But because on 64bit
systems EmuTime::param anyway takes a copy, the bug was not visible there.
Before revision 13164, there was already another copy made of this EmuTime, so
there this bug was hidden.

Bug was trivial to fix once the problem was bisected.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13170 7512d703-071c-0410-b5c2-d2bdd895b7d1