Fixed VC++ windows 64-bit specific bug in HQ resampler asm code
The calling convention on windows is different in 32 and in 64-bit mode. One
(obvious) difference is the way how function parameters are passed. But another
difference is that in 64-bit mode (not in 32-bit mode), the registers xmm6 and
xmm7 should be preserved by an asm routine. The asm routine in
ResampleHQ-x64.asm failed to do this. Once the cause was known, fixing it was
not so difficult. See also
  http://en.wikipedia.org/wiki/X86_calling_conventions#x86-64_calling_conventions

When using gcc inline asm, this is not a problem because there you can specify
exactly which registers are changed or not (so the compiler can automatically
make the routine conform to the ABI). Most likely, when Max originally ported
this routine to windows, it (happened to) work in 64-bit mode. Though
unfortunately these type of bugs often go undetected and only surface when
either changing the caller code or by switching to a better optimizing compiler
(one that actually keeps something useful in the xmm6 and xmm7 registers).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13172 7512d703-071c-0410-b5c2-d2bdd895b7d1