Hold XMLElement objects by value instead of by pointer to heap-allocated object
This main difference is inside the XMLElement class itself. This class has an
array of child XMLElements. Before this patch those children were a vector of
(unique_)pointers to XMLElements. Now the children are stored directly (by
value) in the vector.

The advantage is that there's one less heap-allocation per XMLElement. There
are about 100 such elements per MSX machine. In the past we also used
XMLElements to load the rom database, for that case this change could make a
big difference, for the machine configs the difference is not that significant
anymore.

A disadvantge of storing the child elements by value, is that when you add a
(big) subtree as a child to some other element, you (at least conceptually)
have to make a copy of that whole subtree. In pre-c++11 this would indeed have
been the case. Though now we have move-semantics and I verified that indeed all
those copies can be optimized to moves. (A copy of a subtree would be O(N),
with N the number of elements in the subtree, but a move is O(1)).

--

We use XMLElements a lot less than we used to. So for memory usage and speed
this patch is not that important anymore. Though I could still measure a small
speed gain for this change in the serialization code: serializing value objects
is slightly cheaper because we don't have to keep track of pointer locations.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13177 7512d703-071c-0410-b5c2-d2bdd895b7d1